requirements.txt
alembic==1.3.0
amqp==2.5.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
billiard==3.6.1.0
celery==4.3.0
Click==7.0
factory-boy==2.12.0
Faker==2.0.3
Flask==1.1.1
flask-accepts==0.14.2
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
kombu==4.6.5
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.2.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.10
text-unidecode==1.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
